fixed sql table cannot write custom exposed type
